Cleanup Windows.old folder
Windows creates the folder Windows.old on a PC running Windows when you upgrade the PC to a new version of the operating system. The Windows.old folder contains files and folders of the old installation of Windows, and the main reason it is there is that it is used to restore the old system if the user wants to go back to it. This will clean-up files from a previous Windows installation using the Windows command prompt. Recommendation: Cleanup if your system runs smoothly.

Script is based upon the guide on https://www.ghacks.net/2017/07/12/remove-the-windows-old-folder-manually